requirements.txt
openpyxl
pandas 
python-pptx